requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
langchain-classic
pydantic
fpdf==1.7.2